requirements.txt
attrs==19.3.0
boto==2.49.0
boto3==1.13.22
botocore==1.16.22
bpemb==0.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
Deprecated==1.2.10
docutils==0.15.2
filelock==3.0.12
flair==0.5
Flask==1.1.2
future==0.18.2
gensim==3.8.3
gunicorn==20.0.4
hyperopt==0.2.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
langdetect==1.0.8
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.3.0
mpld3==0.3
networkx==2.4
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
regex==2020.5.14
requests==2.23.0
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
segtok==1.5.10
sentencepiece==0.1.91
six==1.15.0
smart-open==2.0.0
sqlitedict==1.6.0
tabulate==0.8.7
threadpoolctl==2.1.0
tokenizers==0.7.0
https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
tqdm==4.46.1
transformers==2.11.0
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
